@@ -1,11 +1,12 @@
 {
   "name": "platformio-ide",
   "main": "./lib/main",
-  "version": "2.0.0-beta.2",
-  "description": "Official PlatformIO IDE for Atom. PlatformIO is an open source ecosystem for IoT development. Cross-platform build system and library manager. Continuous and IDE integration. Arduino and ARM mbed compatible.",
+  "version": "2.0.0-beta.4",
+  "description": "Official PlatformIO IDE for Atom. PlatformIO is an open source ecosystem for IoT development. Cross-platform IDE and unified debugger. Remote unit testing and firmware updates.",
   "keywords": [
     "PlatformIO",
     "IoT",
+    "debug",
     "embedded",
     "arduino",
     "mbed",
@@ -28,9 +29,9 @@
   "bugs": {
     "url": "https://github.com/platformio/platformio-atom-ide/issues"
   },
-  "license": "https://pioplus.com/license.html",
+  "license": "Apache-2.0",
   "engines": {
-    "atom": ">=1.12.2 <2.0.0"
+    "atom": ">=1.13.0 <2.0.0"
   },
   "activationHooks": [
     "core:loaded-shell-environment"
@@ -58,28 +59,35 @@
       "versions": {
         "^1.0.0": "consumeBusy"
       }
+    },
+    "platformioIDEDebugger": {
+      "versions": {
+        "^1.0.0": "consumeDebugger"
+      }
     }
   },
   "dependencies": {
     "atom-select-list": "^0.2.0",
-    "command-join": "^1.0.1",
+    "command-join": "^1.1.1",
     "flexboxgrid": "^6.3.1",
     "fs-extra": "^2.1.2",
     "fs-plus": "^3.0.0",
     "fuzzaldrin-plus": "^0.4.1",
     "getmac": "^1.2.1",
     "glob": "^7.1.1",
-    "highlight.js": "^9.10.0",
+    "highlight.js": "^9.11.0",
     "humanize": "^0.0.9",
     "ini": "^1.3.4",
     "klaw": "^1.3.1",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-redux": "^5.0.3",
-    "react-router": "^4.0.0",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-redux": "^5.0.4",
+    "react-router": "^4.1.1",
+    "react-simple-dropdown": "^3.0.0",
     "reactable": "^0.14.1",
     "redux": "^3.6.0",
-    "redux-saga": "^0.14.3",
+    "redux-saga": "^0.15.3",
     "request": "^2.81.0",
     "semver": "^5.3.0",
     "tar": "^2.2.1",
@@ -89,9 +97,9 @@
   "devDependencies": {
     "babel-eslint": "^7.2.0",
     "esformatter": "^0.10.0",
-    "esformatter-jsx": "^7.4.0",
+    "esformatter-jsx": "^8.0.0",
     "eslint": "^3.18.0",
-    "eslint-plugin-react": "^6.10.0",
+    "eslint-plugin-react": "^7.0.0",
     "eslint-plugin-redux-saga": "^0.3.0",
     "redux-logger": "^3.0.0"
   }
