@@ -9,16 +9,19 @@
  */
 
 import { Route, Switch } from 'react-router';
-import { getActiveRoute, goTo } from './helpers';
 
+import PropTypes from 'prop-types';
 import React from 'react';
+import UserStatusContainer from '../account/containers/user-status';
+import { getActiveRoute } from './helpers';
+import { goTo } from '../core/helpers';
 import routes from './routes';
 
 
 export default class HomeApp extends React.Component {
 
   static contextTypes = {
-    router: React.PropTypes.object.isRequired
+    router: PropTypes.object.isRequired
   }
 
   render() {
@@ -28,7 +31,7 @@ export default class HomeApp extends React.Component {
         <div className='menu'>
           <ul className='nav nav-pills nav-stacked'>
             { routes.map(item => item.label && (
-                <li onClick={ () => goTo(router.history, item.path) } className={ getActiveRoute(routes, router.route.location) === item ? 'selected' : '' }>
+                <li onClick={ () => goTo(router.history, item.path) } key={ item.path } className={ getActiveRoute(routes, router.route.location) === item ? 'selected' : '' }>
                   <a className={ item.hasOwnProperty('icon') ? 'icon icon-' + item.icon : 'no-icon' }>
                     { item.label }
                   </a>
@@ -52,14 +55,17 @@ export default class HomeApp extends React.Component {
                   ) }
               </div>
               <div className='col-xs text-right'>
-                <span className='icon icon-person'>Developer Preview</span>
+                <UserStatusContainer goTo={ (location) => goTo(router.history, location) } />
               </div>
             </div>
           </div>
           <div className='main-container'>
             <Switch>
               { routes.slice(0).reverse().map(item => (
-                  <Route path={ item.path } exact={ item.exact } component={ item.component } />
+                  <Route path={ item.path }
+                    key={ item.path }
+                    exact={ item.exact }
+                    component={ item.component } />
                 )) }
             </Switch>
           </div>
