@@ -11,6 +11,7 @@
 import * as utils from '../../utils';
 
 import { PKG_BASE_DIR } from '../../config';
+import PropTypes from 'prop-types';
 import React from 'react';
 import fs from 'fs-plus';
 import hljs from 'highlight.js';
@@ -20,9 +21,9 @@ import path from 'path';
 export default class CodeHighlight extends React.Component {
 
   static propTypes = {
-    lang: React.PropTypes.string,
-    url: React.PropTypes.string,
-    content: React.PropTypes.string
+    lang: PropTypes.string,
+    url: PropTypes.string,
+    content: PropTypes.string
   }
 
   componentDidMount() {
