@@ -9,19 +9,21 @@
  */
 
 import { Route, Switch } from 'react-router';
-import { getActiveRoute, goTo } from '../helpers';
 
+import PropTypes from 'prop-types';
 import React from 'react';
+import { getActiveRoute } from '../helpers';
+import { goTo } from '../../core/helpers';
 
 
 export default class SubPages extends React.Component {
 
   static contextTypes = {
-    router: React.PropTypes.object.isRequired
+    router: PropTypes.object.isRequired
   };
 
   static propTypes = {
-    routes: React.PropTypes.arrayOf(React.PropTypes.object).isRequired
+    routes: PropTypes.arrayOf(PropTypes.object).isRequired
   }
 
   render() {
@@ -29,17 +31,23 @@ export default class SubPages extends React.Component {
       <div>
         <div className='page-submenu btn-group btn-group-lg'>
           { this.props.routes.map(item => item.label && (
-              <button onClick={ () => goTo(this.context.router.history, item.path, item.pathState) } className={ ['btn', 'icon', `icon-${item.icon}`, getActiveRoute(this.props.routes, this.context.router.route.location) === item ? 'selected' : ''].join(' ') }>
+              <button onClick={ () => goTo(this.context.router.history, item.path, item.pathState) } key={ item.path } className={ ['btn', 'icon', `icon-${item.icon}`, getActiveRoute(this.props.routes, this.context.router.route.location) === item ? 'selected' : ''].join(' ') }>
                 { item.label }
               </button>
             )) }
         </div>
         <Switch>
           { this.props.routes.slice(0).reverse().map(item => typeof item.component === 'function' ? (
-            <Route path={ item.path } exact={ item.exact } component={ item.component } />
-          ) : (
-            <Route path={ item.path } exact={ item.exact } render={ () => item.component } />
-          ))}
+              <Route path={ item.path }
+                key={ item.path }
+                exact={ item.exact }
+                component={ item.component } />
+              ) : (
+              <Route path={ item.path }
+                key={ item.path }
+                exact={ item.exact }
+                render={ () => item.component } />
+              )) }
         </Switch>
       </div>
     );
