@@ -8,8 +8,9 @@
  * the root directory of this source tree.
  */
 
-import BaseModal from '../base-modal';
+import BaseModal from '../core/base-modal';
 import BaseStage from './stages/base';
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
@@ -24,7 +25,7 @@ export default class InstallerProgressModal extends BaseModal {
 class ModalComponent extends React.Component {
 
   static propTypes = {
-    stages: React.PropTypes.arrayOf(React.PropTypes.object).isRequired
+    stages: PropTypes.arrayOf(PropTypes.object).isRequired
   }
 
   getStatusClass(status) {
@@ -66,7 +67,7 @@ class ModalComponent extends React.Component {
         <div className='select-list'>
           <ol className='list-group'>
             { this.props.stages.map(stage => (
-                <li>
+                <li key={ stage.name }>
                   <div className={ this.getStatusClass(stage.status) }></div>
                   <div className='icon icon-chevron-right'>
                     { stage.name }
