@@ -9,6 +9,7 @@
  */
 
 import CodeHighlight from '../../home/components/code-highlight';
+import PropTypes from 'prop-types';
 import React from 'react';
 import path from 'path';
 
@@ -16,7 +17,7 @@ import path from 'path';
 export default class LibraryDetailExamples extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(React.PropTypes.string).isRequired
+    items: PropTypes.arrayOf(PropTypes.string).isRequired
   }
 
   componentWillMount() {
@@ -62,7 +63,7 @@ export default class LibraryDetailExamples extends React.Component {
 class ExampleFiles extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(React.PropTypes.string).isRequired
+    items: PropTypes.arrayOf(PropTypes.string).isRequired
   }
 
   onDidToggle(event, index) {
