@@ -8,13 +8,14 @@
  * the root directory of this source tree.
  */
 
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class LibraryDetailHeaders extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(React.PropTypes.string).isRequired
+    items: PropTypes.arrayOf(PropTypes.string).isRequired
   }
 
   render() {
