@@ -10,6 +10,7 @@
 
 import * as utils from '../../utils';
 
+import PropTypes from 'prop-types';
 import React from 'react';
 import humanize from 'humanize';
 
@@ -17,22 +18,22 @@ import humanize from 'humanize';
 export default class LibraryDetailMain extends React.Component {
 
   static propTypes = {
-    data: React.PropTypes.shape({
-      id: React.PropTypes.number,
-      name: React.PropTypes.string.isRequired,
-      description: React.PropTypes.string.isRequired,
-      homepage: React.PropTypes.string,
-      repository: React.PropTypes.string,
-      keywords: React.PropTypes.array.isRequired,
-      authors: React.PropTypes.array.isRequired,
-      platforms: React.PropTypes.array,
-      frameworks: React.PropTypes.array,
-      dlstats: React.PropTypes.object,
-      version: React.PropTypes.object,
-      __src_url: React.PropTypes.string,
-      __pkg_dir: React.PropTypes.string
+    data: PropTypes.shape({
+      id: PropTypes.number,
+      name: PropTypes.string.isRequired,
+      description: PropTypes.string.isRequired,
+      homepage: PropTypes.string,
+      repository: PropTypes.string,
+      keywords: PropTypes.array.isRequired,
+      authors: PropTypes.array.isRequired,
+      platforms: PropTypes.array,
+      frameworks: PropTypes.array,
+      dlstats: PropTypes.object,
+      version: PropTypes.object,
+      __src_url: PropTypes.string,
+      __pkg_dir: PropTypes.string
     }),
-    searchLibrary: React.PropTypes.func.isRequired
+    searchLibrary: PropTypes.func.isRequired
   }
 
   getAuthorNames() {
