@@ -9,24 +9,25 @@
  */
 
 import { CompositeDisposable } from 'atom';
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class LibrarySearchCard extends React.Component {
 
   static propTypes = {
-    item: React.PropTypes.shape({
-      id: React.PropTypes.number.isRequired,
-      name: React.PropTypes.string.isRequired,
-      description: React.PropTypes.string.isRequired,
-      examplenums: React.PropTypes.number.isRequired,
-      dlmonth: React.PropTypes.number.isRequired,
-      keywords: React.PropTypes.arrayOf(React.PropTypes.string).isRequired,
-      authornames: React.PropTypes.arrayOf(React.PropTypes.string).isRequired
+    item: PropTypes.shape({
+      id: PropTypes.number.isRequired,
+      name: PropTypes.string.isRequired,
+      description: PropTypes.string.isRequired,
+      examplenums: PropTypes.number.isRequired,
+      dlmonth: PropTypes.number.isRequired,
+      keywords: PropTypes.arrayOf(PropTypes.string).isRequired,
+      authornames: PropTypes.arrayOf(PropTypes.string).isRequired
     }),
-    searchLibrary: React.PropTypes.func.isRequired,
-    showLibrary: React.PropTypes.func.isRequired,
-    installLibrary: React.PropTypes.func.isRequired
+    searchLibrary: PropTypes.func.isRequired,
+    showLibrary: PropTypes.func.isRequired,
+    installLibrary: PropTypes.func.isRequired
   }
 
   componentDidMount() {
