@@ -8,14 +8,15 @@
  * the root directory of this source tree.
  */
 
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class LibrarySearchForm extends React.Component {
 
   static propTypes = {
-    searchLibrary: React.PropTypes.func.isRequired,
-    defaultSearch: React.PropTypes.string
+    searchLibrary: PropTypes.func.isRequired,
+    defaultSearch: PropTypes.string
   }
 
   componentWillReceiveProps(nextProps) {
