@@ -8,23 +8,24 @@
  * the root directory of this source tree.
  */
 
+import PropTypes from 'prop-types';
 import React from 'react';
 import humanize from 'humanize';
 
 export default class LibraryStats extends React.Component {
 
   static propTypes = {
-    data: React.PropTypes.shape({
-      added: React.PropTypes.array.isRequired,
-      dlday: React.PropTypes.array.isRequired,
-      dlmonth: React.PropTypes.array.isRequired,
-      dlweek: React.PropTypes.array.isRequired,
-      lastkeywords: React.PropTypes.array.isRequired,
-      topkeywords: React.PropTypes.array.isRequired,
-      updated: React.PropTypes.array.isRequired
+    data: PropTypes.shape({
+      added: PropTypes.array.isRequired,
+      dlday: PropTypes.array.isRequired,
+      dlmonth: PropTypes.array.isRequired,
+      dlweek: PropTypes.array.isRequired,
+      lastkeywords: PropTypes.array.isRequired,
+      topkeywords: PropTypes.array.isRequired,
+      updated: PropTypes.array.isRequired
     }),
-    searchLibrary: React.PropTypes.func.isRequired,
-    showLibrary: React.PropTypes.func.isRequired
+    searchLibrary: PropTypes.func.isRequired,
+    showLibrary: PropTypes.func.isRequired
   }
 
   getKeywordBtnClassSize(index) {
@@ -51,7 +52,7 @@ export default class LibraryStats extends React.Component {
             <h3 className='block section-heading icon icon-sync'>Updated</h3>
             <ul>
               { this.props.data.updated.map((item) => (
-                  <li>
+                  <li key={ item.name }>
                     <a onClick={ () => this.props.showLibrary(item.id) }>
                       { item.name }
                     </a>
@@ -64,7 +65,7 @@ export default class LibraryStats extends React.Component {
             <h3 className='block section-heading icon icon-file-add'>Added</h3>
             <ul>
               { this.props.data.added.map((item) => (
-                  <li>
+                  <li key={ item.name }>
                     <a onClick={ () => this.props.showLibrary(item.id) }>
                       { item.name }
                     </a>
@@ -77,7 +78,7 @@ export default class LibraryStats extends React.Component {
             <h3 className='block section-heading icon icon-tag'>Keywords</h3>
             <ul className='last-keywords'>
               { this.props.data.lastkeywords.map((name) => (
-                  <li>
+                  <li key={ name }>
                     <button onClick={ () => this.onDidKeywordSearch(name) } className='btn btn-sm icon icon-tag inline-block-tight'>
                       { name }
                     </button>
@@ -89,18 +90,18 @@ export default class LibraryStats extends React.Component {
         <h2 className='block section-heading icon icon-tag'>Popular Keywords</h2>
         <div className='block inline-buttons top-keywords'>
           { this.props.data.topkeywords.map((name, index) => (
-              <button onClick={ () => this.onDidKeywordSearch(name) } className={ 'btn icon icon-tag inline-block-tight btn-' + this.getKeywordBtnClassSize(index) }>
+              <button key={ name } onClick={ () => this.onDidKeywordSearch(name) } className={ 'btn icon icon-tag inline-block-tight btn-' + this.getKeywordBtnClassSize(index) }>
                 { name }
               </button>
             )) }
         </div>
-        <h2 className='block section-heading icon icon-star'>Featured</h2>
+        <h2 className='block section-heading icon icon-star'>Trending</h2>
         <div className='block row'>
           <div className='col-xs'>
             <h3 className='block section-heading icon icon-triangle-up'>Today</h3>
             <ul>
               { this.props.data.dlday.map((item) => (
-                  <li>
+                  <li key={ item.name }>
                     <a onClick={ () => this.props.showLibrary(item.id) }>
                       { item.name }
                     </a>
@@ -112,7 +113,7 @@ export default class LibraryStats extends React.Component {
             <h3 className='block section-heading icon icon-triangle-up'>Week</h3>
             <ul>
               { this.props.data.dlweek.map((item) => (
-                  <li>
+                  <li key={ item.name }>
                     <a onClick={ () => this.props.showLibrary(item.id) }>
                       { item.name }
                     </a>
@@ -124,7 +125,7 @@ export default class LibraryStats extends React.Component {
             <h3 className='block section-heading icon icon-triangle-up'>Month</h3>
             <ul>
               { this.props.data.dlmonth.map((item) => (
-                  <li>
+                  <li key={ item.name }>
                     <a onClick={ () => this.props.showLibrary(item.id) }>
                       { item.name }
                     </a>
