@@ -9,28 +9,29 @@
  */
 
 import { LibraryStorage } from '../storage';
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class LibraryStorageItem extends React.Component {
 
   static propTypes = {
-    item: React.PropTypes.shape({
-      name: React.PropTypes.string.isRequired,
-      version: React.PropTypes.string.isRequired,
-      versionLatest: React.PropTypes.string,
-      description: React.PropTypes.string,
-      url: React.PropTypes.string,
-      keywords: React.PropTypes.arrayOf(React.PropTypes.string).isRequired,
-      authors: React.PropTypes.arrayOf(React.PropTypes.object),
-      __src_url: React.PropTypes.string.isRequired,
+    item: PropTypes.shape({
+      name: PropTypes.string.isRequired,
+      version: PropTypes.string.isRequired,
+      versionLatest: PropTypes.string,
+      description: PropTypes.string,
+      url: PropTypes.string,
+      keywords: PropTypes.arrayOf(PropTypes.string).isRequired,
+      authors: PropTypes.arrayOf(PropTypes.object),
+      __src_url: PropTypes.string.isRequired,
     }),
-    onShow: React.PropTypes.func.isRequired,
-    onReveal: React.PropTypes.func.isRequired,
-    onUninstall: React.PropTypes.func.isRequired,
-    onUpdate: React.PropTypes.func.isRequired,
-    onSearch: React.PropTypes.func.isRequired,
-    actions: React.PropTypes.number.isRequired
+    onShow: PropTypes.func.isRequired,
+    onReveal: PropTypes.func.isRequired,
+    onUninstall: PropTypes.func.isRequired,
+    onUpdate: PropTypes.func.isRequired,
+    onSearch: PropTypes.func.isRequired,
+    actions: PropTypes.number.isRequired
   }
 
   getAuthorNames() {
@@ -71,7 +72,7 @@ export default class LibraryStorageItem extends React.Component {
         <div className='row bottom-xs'>
           <div className='col-xs-7 inline-buttons'>
             { (this.props.item.keywords || []).map(name => (
-                <button onClick={ (e) => this.onDidKeywordSearch(e, name) } className='btn btn-sm icon icon-tag inline-block-tight'>
+                <button onClick={ (e) => this.onDidKeywordSearch(e, name) } key={ name } className='btn btn-sm icon icon-tag inline-block-tight'>
                   { name }
                 </button>
               )) }
