@@ -12,17 +12,18 @@ import * as utils from '../../utils';
 
 import { LibraryStorage } from '../storage';
 import LibraryStorageItem from './storage-item';
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class LibraryStorageItems extends React.Component {
 
   static propTypes = {
-    item: React.PropTypes.instanceOf(LibraryStorage).isRequired,
-    searchLibrary: React.PropTypes.func.isRequired,
-    showLibrary: React.PropTypes.func.isRequired,
-    uninstallLibrary: React.PropTypes.func.isRequired,
-    updateLibrary: React.PropTypes.func.isRequired
+    item: PropTypes.instanceOf(LibraryStorage).isRequired,
+    searchLibrary: PropTypes.func.isRequired,
+    showLibrary: PropTypes.func.isRequired,
+    uninstallLibrary: PropTypes.func.isRequired,
+    updateLibrary: PropTypes.func.isRequired
   }
 
   static status = {
@@ -101,6 +102,7 @@ export default class LibraryStorageItems extends React.Component {
         <div ref='storageItems'>
           { items.map(item => (
               <LibraryStorageItem item={ item }
+                key={ item.name }
                 onShow={ this.props.showLibrary }
                 onReveal={ (e) => this.onDidReveal(e, item.__pkg_dir) }
                 onUninstall={ (e) => this.onDidUninstallOrUpdateItem(e, item, 'uninstall') }
