@@ -10,21 +10,22 @@
 
 import { LibraryStorage } from '../storage';
 import LibraryStorageItems from './storage-items';
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class LibraryStoragesList extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(
-      React.PropTypes.instanceOf(LibraryStorage).isRequired
+    items: PropTypes.arrayOf(
+      PropTypes.instanceOf(LibraryStorage).isRequired
     ),
-    filterValue: React.PropTypes.string.isRequired,
-    setFilter: React.PropTypes.func.isRequired,
-    searchLibrary: React.PropTypes.func.isRequired,
-    showLibrary: React.PropTypes.func.isRequired,
-    uninstallLibrary: React.PropTypes.func.isRequired,
-    updateLibrary: React.PropTypes.func.isRequired
+    filterValue: PropTypes.string,
+    setFilter: PropTypes.func.isRequired,
+    searchLibrary: PropTypes.func.isRequired,
+    showLibrary: PropTypes.func.isRequired,
+    uninstallLibrary: PropTypes.func.isRequired,
+    updateLibrary: PropTypes.func.isRequired
   }
 
   render() {
@@ -58,8 +59,9 @@ export default class LibraryStoragesList extends React.Component {
             onChange={ (e) => this.props.setFilter(e.target.value) } />
         </div>
         <div className='lib-storages'>
-          { this.props.items.map(item => (
+          { this.props.items.map((item, index) => (
               <LibraryStorageItems item={ item }
+                key={ index }
                 searchLibrary={ this.props.searchLibrary }
                 showLibrary={ this.props.showLibrary }
                 uninstallLibrary={ this.props.uninstallLibrary }
