@@ -15,25 +15,25 @@ import { BUILTIN_INPUT_FILTER_KEY, getBuiltinFilter, getVisibletBuiltinLibs } fr
 import { INPUT_FILTER_DELAY } from '../../config';
 import { LibraryStorage } from '../storage';
 import LibraryStoragesList from '../components/storages-list';
+import PropTypes from 'prop-types';
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { goTo } from '../../home/helpers';
-import { lazyUpdateInputValue } from '../../home/actions';
+import { goTo } from '../../core/helpers';
+import { lazyUpdateInputValue } from '../../core/actions';
 
 
 class LibraryBuiltinPage extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(
-      React.PropTypes.instanceOf(LibraryStorage).isRequired
+    items: PropTypes.arrayOf(
+      PropTypes.instanceOf(LibraryStorage).isRequired
     ),
-    storageFilterValue: React.PropTypes.string.isRequired,
-    setStorageFilter: React.PropTypes.func.isRequired,
-    loadBuiltinLibs: React.PropTypes.func.isRequired,
-    searchLibrary: React.PropTypes.func.isRequired,
-    showLibrary: React.PropTypes.func.isRequired,
-    showInstalledPlatforms: React.PropTypes.func.isRequired
+    filterValue: PropTypes.string,
+    loadBuiltinLibs: PropTypes.func.isRequired,
+    searchLibrary: PropTypes.func.isRequired,
+    showLibrary: PropTypes.func.isRequired,
+    showInstalledPlatforms: PropTypes.func.isRequired
   }
 
   componentWillMount() {
