@@ -9,6 +9,7 @@
  */
 
 import LibraryDetailExamples from '../components/detail-examples';
+import PropTypes from 'prop-types';
 import React from 'react';
 import fs from 'fs-plus';
 import glob from 'glob';
@@ -18,9 +19,9 @@ import path from 'path';
 export default class LibraryDetailExamplesBlock extends React.Component {
 
   static propTypes = {
-    data: React.PropTypes.shape({
-      examples: React.PropTypes.array,
-      __pkg_dir: React.PropTypes.string
+    data: PropTypes.shape({
+      examples: PropTypes.array,
+      __pkg_dir: PropTypes.string
     }).isRequired
   }
 
