@@ -9,6 +9,7 @@
  */
 
 import LibraryDetailHeaders from '../components/detail-headers';
+import PropTypes from 'prop-types';
 import React from 'react';
 import klaw from 'klaw';
 import path from 'path';
@@ -17,9 +18,9 @@ import path from 'path';
 export default class LibraryDetailHeadersBlock extends React.Component {
 
   static propTypes = {
-    data: React.PropTypes.shape({
-      headers: React.PropTypes.arrayOf(React.PropTypes.string),
-      __pkg_dir: React.PropTypes.string
+    data: PropTypes.shape({
+      headers: PropTypes.arrayOf(PropTypes.string),
+      __pkg_dir: PropTypes.string
     }).isRequired
   }
 
