@@ -9,6 +9,7 @@
  */
 
 import CodeHighlight from '../../home/components/code-highlight';
+import PropTypes from 'prop-types';
 import React from 'react';
 import humanize from 'humanize';
 
@@ -16,21 +17,21 @@ import humanize from 'humanize';
 export default class LibraryDetailInstallationBlock extends React.Component {
 
   static propTypes = {
-    data: React.PropTypes.shape({
-      id: React.PropTypes.number.isRequired,
-      name: React.PropTypes.string.isRequired,
-      version: React.PropTypes.object,
-      versions: React.PropTypes.array,
-      platforms: React.PropTypes.arrayOf(React.PropTypes.shape({
-        name: React.PropTypes.string,
-        title: React.PropTypes.string
+    data: PropTypes.shape({
+      id: PropTypes.number.isRequired,
+      name: PropTypes.string.isRequired,
+      version: PropTypes.object,
+      versions: PropTypes.array,
+      platforms: PropTypes.arrayOf(PropTypes.shape({
+        name: PropTypes.string,
+        title: PropTypes.string
       })),
-      frameworks: React.PropTypes.arrayOf(React.PropTypes.shape({
-        name: React.PropTypes.string,
-        title: React.PropTypes.string
+      frameworks: PropTypes.arrayOf(PropTypes.shape({
+        name: PropTypes.string,
+        title: PropTypes.string
       }))
     }).isRequired,
-    installLibrary: React.PropTypes.func.isRequired
+    installLibrary: PropTypes.func.isRequired
   }
 
   onDidVersionInstall(event, version) {
