@@ -11,14 +11,15 @@
 import * as utils from '../../utils';
 
 import CodeHighlight from '../../home/components/code-highlight';
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class LibraryDetailManifestBlock extends React.Component {
 
   static propTypes = {
-    data: React.PropTypes.shape({
-      confurl: React.PropTypes.string
+    data: PropTypes.shape({
+      confurl: PropTypes.string
     }).isRequired
   }
 
