@@ -16,30 +16,31 @@ import LibraryDetailHeadersBlock from '../containers/detail-headers-block';
 import LibraryDetailInstallationBlock from '../containers/detail-installation-block';
 import LibraryDetailMain from '../components/detail-main';
 import LibraryDetailManifestBlock from '../containers/detail-manifest-block';
+import PropTypes from 'prop-types';
 import React from 'react';
 import SubPages from '../../home/components/subpages';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { getLibraryData } from '../selectors';
-import { goTo } from '../../home/helpers';
+import { goTo } from '../../core/helpers';
 
 
 class LibraryDetailPage extends React.Component {
 
   static contextTypes = {
-    router: React.PropTypes.object.isRequired
+    router: PropTypes.object.isRequired
   };
 
   static propTypes = {
-    data: React.PropTypes.object,
-    idOrManifest: React.PropTypes.oneOfType([
-      React.PropTypes.number,
-      React.PropTypes.object
+    data: PropTypes.object,
+    idOrManifest: PropTypes.oneOfType([
+      PropTypes.number,
+      PropTypes.object
     ]),
-    loadLibraryData: React.PropTypes.func.isRequired,
-    installLibrary: React.PropTypes.func.isRequired,
-    searchLibrary: React.PropTypes.func.isRequired,
-    showLibrary: React.PropTypes.func.isRequired
+    loadLibraryData: PropTypes.func.isRequired,
+    installLibrary: PropTypes.func.isRequired,
+    searchLibrary: PropTypes.func.isRequired,
+    showLibrary: PropTypes.func.isRequired
   }
 
   componentWillMount() {
