@@ -8,7 +8,8 @@
  * the root directory of this source tree.
  */
 
-import BaseModal from '../../base-modal';
+import BaseModal from '../../core/base-modal';
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
@@ -23,8 +24,8 @@ export default class LibraryInstallAdvancedModal extends BaseModal {
 class ModalComponent extends React.Component {
 
   static propTypes = {
-    onResolve: React.PropTypes.func.isRequired,
-    onReject: React.PropTypes.func.isRequired
+    onResolve: PropTypes.func.isRequired,
+    onReject: PropTypes.func.isRequired
   }
 
   onDidSubmit() {
