@@ -15,25 +15,26 @@ import { INSTALLED_INPUT_FILTER_KEY, getInstalledFilter, getVisibleInstalledLibs
 import { INPUT_FILTER_DELAY } from '../../config';
 import { LibraryStorage } from '../storage';
 import LibraryStoragesList from '../components/storages-list';
+import PropTypes from 'prop-types';
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { goTo } from '../../home/helpers';
-import { lazyUpdateInputValue } from '../../home/actions';
+import { goTo } from '../../core/helpers';
+import { lazyUpdateInputValue } from '../../core/actions';
 
 
 class LibraryInstalledPage extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(
-      React.PropTypes.instanceOf(LibraryStorage).isRequired
+    items: PropTypes.arrayOf(
+      PropTypes.instanceOf(LibraryStorage).isRequired
     ),
-    filterValue: React.PropTypes.string.isRequired,
-    setFilter: React.PropTypes.func.isRequired,
-    loadInstalledLibs: React.PropTypes.func.isRequired,
-    searchLibrary: React.PropTypes.func.isRequired,
-    showLibrary: React.PropTypes.func.isRequired,
-    uninstallLibrary: React.PropTypes.func.isRequired
+    filterValue: PropTypes.string,
+    setFilter: PropTypes.func.isRequired,
+    loadInstalledLibs: PropTypes.func.isRequired,
+    searchLibrary: PropTypes.func.isRequired,
+    showLibrary: PropTypes.func.isRequired,
+    uninstallLibrary: PropTypes.func.isRequired
   }
 
   componentWillMount() {
