@@ -12,28 +12,29 @@ import * as actions from '../actions';
 
 import LibrarySearchCard from '../components/search-card';
 import LibrarySearchForm from '../components/search-form';
+import PropTypes from 'prop-types';
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { getSearchResult } from '../selectors';
-import { goTo } from '../../home/helpers';
+import { goTo } from '../../core/helpers';
 
 
 class LibrarySearchPage extends React.Component {
 
   static propTypes = {
-    result: React.PropTypes.shape({
-      items: React.PropTypes.array.isRequired,
-      total: React.PropTypes.number.isRequired,
-      page: React.PropTypes.number.isRequired,
-      perpage: React.PropTypes.number.isRequired
+    result: PropTypes.shape({
+      items: PropTypes.array.isRequired,
+      total: PropTypes.number.isRequired,
+      page: PropTypes.number.isRequired,
+      perpage: PropTypes.number.isRequired
     }),
-    searchLibrary: React.PropTypes.func.isRequired,
-    showLibrary: React.PropTypes.func.isRequired,
-    loadSearchResult: React.PropTypes.func.isRequired,
-    installLibrary: React.PropTypes.func.isRequired,
-    searchQuery: React.PropTypes.string.isRequired,
-    searchPage: React.PropTypes.number.isRequired
+    searchLibrary: PropTypes.func.isRequired,
+    showLibrary: PropTypes.func.isRequired,
+    loadSearchResult: PropTypes.func.isRequired,
+    installLibrary: PropTypes.func.isRequired,
+    searchQuery: PropTypes.string.isRequired,
+    searchPage: PropTypes.number.isRequired
   }
 
   static STATUS = {
