@@ -14,21 +14,22 @@ import * as utils from '../../utils';
 import LibraryInstallAdvancedModal from './install-advanced-modal';
 import LibrarySearchForm from '../components/search-form';
 import LibraryStats from '../components/stats';
+import PropTypes from 'prop-types';
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { getStats } from '../selectors';
-import { goTo } from '../../home/helpers';
+import { goTo } from '../../core/helpers';
 
 
 class LibraryStatsPage extends React.Component {
 
   static propTypes = {
-    data: React.PropTypes.object,
-    loadStats: React.PropTypes.func.isRequired,
-    installLibrary: React.PropTypes.func.isRequired,
-    searchLibrary: React.PropTypes.func.isRequired,
-    showLibrary: React.PropTypes.func.isRequired
+    data: PropTypes.object,
+    loadStats: PropTypes.func.isRequired,
+    installLibrary: PropTypes.func.isRequired,
+    searchLibrary: PropTypes.func.isRequired,
+    showLibrary: PropTypes.func.isRequired
   }
 
   componentWillMount() {
