@@ -9,6 +9,7 @@
  */
 
 import { INPUT_FILTER_DELAY } from '../../config';
+import PropTypes from 'prop-types';
 import React from 'react';
 import Reactable from 'reactable';
 import humanize from 'humanize';
@@ -17,31 +18,31 @@ import humanize from 'humanize';
 export default class Boards extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(React.PropTypes.shape({
-      id: React.PropTypes.number.isRequired,
-      name: React.PropTypes.string.isRequired,
-      url: React.PropTypes.string.isRequired,
-      vendor: React.PropTypes.string.isRequired,
-      platform: React.PropTypes.shape({
-        name: React.PropTypes.string.isRequired,
-        title: React.PropTypes.string.isRequired,
+    items: PropTypes.arrayOf(PropTypes.shape({
+      id: PropTypes.number.isRequired,
+      name: PropTypes.string.isRequired,
+      url: PropTypes.string.isRequired,
+      vendor: PropTypes.string.isRequired,
+      platform: PropTypes.shape({
+        name: PropTypes.string.isRequired,
+        title: PropTypes.string.isRequired,
       }).isRequired,
-      frameworks: React.PropTypes.arrayOf(React.PropTypes.shape({
-        name: React.PropTypes.string.isRequired,
-        title: React.PropTypes.string.isRequired,
+      frameworks: PropTypes.arrayOf(PropTypes.shape({
+        name: PropTypes.string.isRequired,
+        title: PropTypes.string.isRequired,
       })).isRequired,
-      mcu: React.PropTypes.string.isRequired,
-      fcpu: React.PropTypes.number.isRequired,
-      rom: React.PropTypes.number.isRequired,
-      ram: React.PropTypes.number.isRequired
+      mcu: PropTypes.string.isRequired,
+      fcpu: PropTypes.number.isRequired,
+      rom: PropTypes.number.isRequired,
+      ram: PropTypes.number.isRequired
     })),
-    header: React.PropTypes.string,
-    headerSize: React.PropTypes.number,
-    itemsPerPage: React.PropTypes.number,
-    defaultFilter: React.PropTypes.string,
-    onFilter: React.PropTypes.func,
-    showPlatform: React.PropTypes.func.isRequired,
-    showFramework: React.PropTypes.func.isRequired
+    header: PropTypes.string,
+    headerSize: PropTypes.number,
+    itemsPerPage: PropTypes.number,
+    defaultFilter: PropTypes.string,
+    onFilter: PropTypes.func,
+    showPlatform: PropTypes.func.isRequired,
+    showFramework: PropTypes.func.isRequired
   }
 
   static defaultProps ={
@@ -60,6 +61,29 @@ export default class Boards extends React.Component {
 
   humanizeMemorySize(size) {
     return humanize.filesize(size, 1024, size % 1024 === 0 || size < 1024 ? 0 : 1);
+  }
+
+  humanizeDebug(data) {
+    if (!data || !data.tools) {
+      return;
+    }
+    const tools = [];
+    Object.keys(data.tools).forEach(key => {
+      const options = data.tools[key];
+      const attrs = [];
+      if (options.default) {
+        attrs.push('default');
+      }
+      if (options.onboard) {
+        attrs.push('on-board');
+      }
+      if (attrs.length) {
+        tools.push(`${key} (${attrs.join(', ')})`);
+      } else {
+        tools.push(key);
+      }
+    });
+    return tools.join(', ');
   }
 
   onFilter(value) {
@@ -92,6 +116,9 @@ export default class Boards extends React.Component {
           ) }
         <div className='text'>
           <span className='icon icon-question'></span>You can use <span className='icon icon-clippy'></span> icon to copy Board ID to clipboard. Please read more how to <a href='http://docs.platformio.org/page/platforms/creating_board.html'>create a custom board</a>. When the table columns can not fit on screen, please scroll horizontally.
+        </div>
+        <div className='text'>
+          <span className='icon icon-bug'></span><a href='http://docs.platformio.org/page/plus/debugging.html'>PIO Unified Debugger</a>: mouse over the icon to list available debugging tools or check which are already on a board.
         </div>
         <br />
         <div className='native-key-bindings' tabIndex='-1'>
@@ -140,9 +167,6 @@ export default class Boards extends React.Component {
           className='native-key-bindings table table-hover reacttable'
           tabIndex='-1'>
           <Thead>
-            <Th column='id'>
-              ID
-            </Th>
             <Th column='name'>
               Name
             </Th>
@@ -151,6 +175,9 @@ export default class Boards extends React.Component {
             </Th>
             <Th column='frameworks'>
               Frameworks
+            </Th>
+            <Th column='debug'>
+              Debug
             </Th>
             <Th column='mcu'>
               MCU
@@ -166,14 +193,14 @@ export default class Boards extends React.Component {
             </Th>
           </Thead>
           { items.map(item => (
-              <Tr>
-                <Td column='id' value={ item.id }>
-                  <span><a onClick={ () => atom.clipboard.write(item.id) }><span title='Copy to clipboard' className='icon icon-clippy'></span></a> <span className='inline-block highlight'>{ item.id }</span></span>
-                </Td>
+              <Tr key={ item.name }>
                 <Td column='name' value={ item.name }>
-                  <a href={ item.url }>
-                    { item.name }
-                  </a>
+                  <span>
+                    <a onClick={ () => atom.clipboard.write(item.id) } title={ `Copy to clipboard: ${item.id}` }><span className='icon icon-clippy'></span></a>
+                    <a href={ item.url }>
+                      { item.name }
+                    </a>
+                  </span>
                 </Td>
                 <Td column='platform' value={ item.platform.name }>
                   <a onClick={ () => this.props.showPlatform(item.platform.name) }>
@@ -182,10 +209,13 @@ export default class Boards extends React.Component {
                 </Td>
                 <Td column='frameworks' value={ item.frameworks.map(f => f.name).join() } className='inline-anchors'>
                   <span>{ item.frameworks.map(framework => (
-                            <a onClick={ () => this.props.showFramework(framework.name) }>
+                            <a onClick={ () => this.props.showFramework(framework.name) } key={ framework.title }>
                               { framework.title }
                             </a>
                           )) }</span>
+                </Td>
+                <Td column='debug' value={ item.debug ? -1 : 1 }>
+                  <span>{ item.debug && <a href='http://docs.platformio.org/page/plus/debugging.html'><span className='icon icon-bug' title={ this.humanizeDebug(item.debug) }></span></a> }</span>
                 </Td>
                 <Td column='mcu' data={ item.mcu }></Td>
                 <Td column='fcpu' value={ item.fcpu } className='text-nowrap'>
