@@ -8,21 +8,22 @@
  * the root directory of this source tree.
  */
 
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class FrameworkDetailPage extends React.Component {
 
   static propTypes = {
-    data: React.PropTypes.shape({
-      name: React.PropTypes.string.isRequired,
-      title: React.PropTypes.string.isRequired,
-      description: React.PropTypes.string.isRequired,
-      homepage: React.PropTypes.string.isRequired,
-      url: React.PropTypes.string,
-      platforms: React.PropTypes.arrayOf(React.PropTypes.object)
+    data: PropTypes.shape({
+      name: PropTypes.string.isRequired,
+      title: PropTypes.string.isRequired,
+      description: PropTypes.string.isRequired,
+      homepage: PropTypes.string.isRequired,
+      url: PropTypes.string,
+      platforms: PropTypes.arrayOf(PropTypes.object)
     }),
-    showPlatform: React.PropTypes.func.isRequired
+    showPlatform: PropTypes.func.isRequired
   }
 
   render() {
