@@ -8,19 +8,20 @@
  * the root directory of this source tree.
  */
 
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class FrameworksList extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(
-      React.PropTypes.object.isRequired
+    items: PropTypes.arrayOf(
+      PropTypes.object.isRequired
     ),
-    filterValue: React.PropTypes.string.isRequired,
-    setFilter: React.PropTypes.func.isRequired,
-    showPlatform: React.PropTypes.func.isRequired,
-    showFramework: React.PropTypes.func.isRequired
+    filterValue: PropTypes.string.isRequired,
+    setFilter: PropTypes.func.isRequired,
+    showPlatform: PropTypes.func.isRequired,
+    showFramework: PropTypes.func.isRequired
   }
 
   onDidShow(event, name) {
