@@ -9,28 +9,30 @@
  */
 
 import * as utils from '../../utils';
+
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class PlatformCard extends React.Component {
 
   static propTypes = {
-    item: React.PropTypes.shape({
-      name: React.PropTypes.string.isRequired,
-      title: React.PropTypes.string.isRequired,
-      description: React.PropTypes.string.isRequired,
-      version: React.PropTypes.object,
-      versionLatest: React.PropTypes.object,
-      frameworks: React.PropTypes.array,
-      __src_url: React.PropTypes.string,
-      __pkg_dir: React.PropTypes.string
+    item: PropTypes.shape({
+      name: PropTypes.string.isRequired,
+      title: PropTypes.string.isRequired,
+      description: PropTypes.string.isRequired,
+      version: PropTypes.object,
+      versionLatest: PropTypes.object,
+      frameworks: PropTypes.array,
+      __src_url: PropTypes.string,
+      __pkg_dir: PropTypes.string
     }),
-    actions: React.PropTypes.arrayOf(React.PropTypes.string).isRequired,
-    showPlatform: React.PropTypes.func.isRequired,
-    showFramework: React.PropTypes.func.isRequired,
-    installPlatform: React.PropTypes.func.isRequired,
-    uninstallPlatform: React.PropTypes.func.isRequired,
-    updatePlatform: React.PropTypes.func.isRequired
+    actions: PropTypes.arrayOf(PropTypes.string).isRequired,
+    showPlatform: PropTypes.func.isRequired,
+    showFramework: PropTypes.func.isRequired,
+    installPlatform: PropTypes.func.isRequired,
+    uninstallPlatform: PropTypes.func.isRequired,
+    updatePlatform: PropTypes.func.isRequired
   }
 
   onDidReveal(event) {
@@ -90,7 +92,7 @@ export default class PlatformCard extends React.Component {
         <div className='row bottom-xs'>
           <div className='col-xs-7 inline-buttons'>
             { (this.props.item.frameworks || []).map(item => (
-                <button onClick={ (e) => this.onDidFramework(e, item.name) } className='btn btn-sm icon icon-gear inline-block-tight'>
+                <button onClick={ (e) => this.onDidFramework(e, item.name) } key={ item.title } className='btn btn-sm icon icon-gear inline-block-tight'>
                   { item.title }
                 </button>
               )) }
