@@ -10,28 +10,29 @@
 
 import * as utils from '../../utils';
 
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class PlatformMain extends React.Component {
 
   static propTypes = {
-    data: React.PropTypes.shape({
-      name: React.PropTypes.string.isRequired,
-      title: React.PropTypes.string.isRequired,
-      description: React.PropTypes.string.isRequired,
-      homepage: React.PropTypes.string.isRequired,
-      repository: React.PropTypes.string,
-      url: React.PropTypes.string,
-      license: React.PropTypes.string,
-      version: React.PropTypes.string,
-      versions: React.PropTypes.arrayOf(React.PropTypes.string),
-      frameworks: React.PropTypes.arrayOf(React.PropTypes.object),
-      __src_url: React.PropTypes.string,
-      __pkg_dir: React.PropTypes.string
+    data: PropTypes.shape({
+      name: PropTypes.string.isRequired,
+      title: PropTypes.string.isRequired,
+      description: PropTypes.string.isRequired,
+      homepage: PropTypes.string.isRequired,
+      repository: PropTypes.string,
+      url: PropTypes.string,
+      license: PropTypes.string,
+      version: PropTypes.string,
+      versions: PropTypes.arrayOf(PropTypes.string),
+      frameworks: PropTypes.arrayOf(PropTypes.object),
+      __src_url: PropTypes.string,
+      __pkg_dir: PropTypes.string
     }),
-    showFramework: React.PropTypes.func.isRequired,
-    installPlatform: React.PropTypes.func.isRequired
+    showFramework: PropTypes.func.isRequired,
+    installPlatform: PropTypes.func.isRequired
   }
 
   onDidReveal(dir) {
