@@ -8,21 +8,22 @@
  * the root directory of this source tree.
  */
 
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class PlatformDetailPackages extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(React.PropTypes.shape({
-      name: React.PropTypes.string.isRequired,
-      type: React.PropTypes.string.isRequired,
-      description: React.PropTypes.string.isRequired,
-      requirements: React.PropTypes.string.isRequired,
-      version: React.PropTypes.string.isRequired,
-      originalVersion: React.PropTypes.string,
-      url: React.PropTypes.string,
-      optional: React.PropTypes.string,
+    items: PropTypes.arrayOf(PropTypes.shape({
+      name: PropTypes.string.isRequired,
+      type: PropTypes.string.isRequired,
+      description: PropTypes.string.isRequired,
+      requirements: PropTypes.string.isRequired,
+      version: PropTypes.string.isRequired,
+      originalVersion: PropTypes.string,
+      url: PropTypes.string,
+      optional: PropTypes.string,
     }))
   }
 
