@@ -9,23 +9,24 @@
  */
 
 import PlatformCard from './platform-card';
+import PropTypes from 'prop-types';
 import React from 'react';
 
 
 export default class PlatformsList extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(
-      React.PropTypes.object.isRequired
+    items: PropTypes.arrayOf(
+      PropTypes.object.isRequired
     ),
-    filterValue: React.PropTypes.string.isRequired,
-    setFilter: React.PropTypes.func.isRequired,
-    actions: React.PropTypes.arrayOf(React.PropTypes.string).isRequired,
-    showPlatform: React.PropTypes.func.isRequired,
-    showFramework: React.PropTypes.func.isRequired,
-    installPlatform: React.PropTypes.func.isRequired,
-    uninstallPlatform: React.PropTypes.func.isRequired,
-    updatePlatform: React.PropTypes.func.isRequired
+    filterValue: PropTypes.string,
+    setFilter: PropTypes.func.isRequired,
+    actions: PropTypes.arrayOf(PropTypes.string).isRequired,
+    showPlatform: PropTypes.func.isRequired,
+    showFramework: PropTypes.func.isRequired,
+    installPlatform: PropTypes.func.isRequired,
+    uninstallPlatform: PropTypes.func.isRequired,
+    updatePlatform: PropTypes.func.isRequired
   }
 
   render() {
@@ -57,16 +58,16 @@ export default class PlatformsList extends React.Component {
           </ul>
         }
         <br />
-        { this.props.items.map(item => (
-          <PlatformCard
-            item={ item }
-            actions={ this.props.actions }
-            showPlatform={ this.props.showPlatform }
-            showFramework={ this.props.showFramework }
-            installPlatform={ this.props.installPlatform }
-            uninstallPlatform={ this.props.uninstallPlatform }
-            updatePlatform={ this.props.updatePlatform } />
-        ))}
+        { this.props.items.map((item, index) => (
+            <PlatformCard item={ item }
+              key={ index }
+              actions={ this.props.actions }
+              showPlatform={ this.props.showPlatform }
+              showFramework={ this.props.showFramework }
+              installPlatform={ this.props.installPlatform }
+              uninstallPlatform={ this.props.uninstallPlatform }
+              updatePlatform={ this.props.updatePlatform } />
+          )) }
       </div>
     );
   }
