@@ -13,24 +13,25 @@ import { BOARDS_INPUT_FILTER_KEY, getBoardsFilter, getVisibleBoards } from '../s
 
 import Boards from '../components/boards';
 import { INPUT_FILTER_DELAY } from '../../config';
+import PropTypes from 'prop-types';
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { goTo } from '../../home/helpers';
-import { lazyUpdateInputValue } from '../../home/actions';
+import { goTo } from '../../core/helpers';
+import { lazyUpdateInputValue } from '../../core/actions';
 
 
 class BoardsPage extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(React.PropTypes.object),
-    storageFilterValue: React.PropTypes.string.isRequired,
-    setStorageFilter: React.PropTypes.func.isRequired,
-    filterValue: React.PropTypes.string.isRequired,
-    setFilter: React.PropTypes.func.isRequired,
-    loadBoards: React.PropTypes.func.isRequired,
-    showPlatform: React.PropTypes.func.isRequired,
-    showFramework: React.PropTypes.func.isRequired
+    items: PropTypes.arrayOf(PropTypes.object),
+    storageFilterValue: PropTypes.string.isRequired,
+    setStorageFilter: PropTypes.func.isRequired,
+    filterValue: PropTypes.string.isRequired,
+    setFilter: PropTypes.func.isRequired,
+    loadBoards: PropTypes.func.isRequired,
+    showPlatform: PropTypes.func.isRequired,
+    showFramework: PropTypes.func.isRequired
   }
 
   componentWillMount() {
