@@ -15,25 +15,26 @@ import { EMBEDDED_INPUT_FILTER_KEY, getEmbeddedFilter, getVisibleEmbeddedPlatfor
 import { INPUT_FILTER_DELAY } from '../../config';
 import PlatformInstallAdvancedModal from './install-advanced-modal';
 import PlatformsList from '../components/platforms-list';
+import PropTypes from 'prop-types';
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { goTo } from '../../home/helpers';
-import { lazyUpdateInputValue } from '../../home/actions';
+import { goTo } from '../../core/helpers';
+import { lazyUpdateInputValue } from '../../core/actions';
 
 
 class PlatformEmbeddedPage extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(
-      React.PropTypes.object.isRequired
+    items: PropTypes.arrayOf(
+      PropTypes.object.isRequired
     ),
-    filterValue: React.PropTypes.string.isRequired,
-    setFilter: React.PropTypes.func.isRequired,
-    loadRegistryPlatforms: React.PropTypes.func.isRequired,
-    showPlatform: React.PropTypes.func.isRequired,
-    showFramework: React.PropTypes.func.isRequired,
-    installPlatform: React.PropTypes.func.isRequired
+    filterValue: PropTypes.string,
+    setFilter: PropTypes.func.isRequired,
+    loadRegistryPlatforms: PropTypes.func.isRequired,
+    showPlatform: PropTypes.func.isRequired,
+    showFramework: PropTypes.func.isRequired,
+    installPlatform: PropTypes.func.isRequired
   }
 
   componentWillMount() {
