@@ -12,21 +12,22 @@ import * as actions from '../actions';
 
 import Boards from '../components/boards';
 import FrameworkDetailMain from '../components/framework-main';
+import PropTypes from 'prop-types';
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { getFrameworkData } from '../selectors';
-import { goTo } from '../../home/helpers';
+import { goTo } from '../../core/helpers';
 
 
 class FrameworkDetailPage extends React.Component {
 
   static propTypes = {
-    name: React.PropTypes.string.isRequired,
-    data: React.PropTypes.object,
-    loadFrameworkData: React.PropTypes.func.isRequired,
-    showPlatform: React.PropTypes.func.isRequired,
-    showFramework: React.PropTypes.func.isRequired
+    name: PropTypes.string.isRequired,
+    data: PropTypes.object,
+    loadFrameworkData: PropTypes.func.isRequired,
+    showPlatform: PropTypes.func.isRequired,
+    showFramework: PropTypes.func.isRequired
   }
 
   componentWillMount() {
