@@ -13,22 +13,23 @@ import * as actions from '../actions';
 import Boards from '../components/boards';
 import PlatformDetailMain from '../components/platform-main';
 import PlatformDetailPackages from '../components/platform-packages';
+import PropTypes from 'prop-types';
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { getPlatformData } from '../selectors';
-import { goTo } from '../../home/helpers';
+import { goTo } from '../../core/helpers';
 
 
 class PlatformDetailPage extends React.Component {
 
   static propTypes = {
-    name: React.PropTypes.string.isRequired,
-    data: React.PropTypes.object,
-    loadPlatformData: React.PropTypes.func.isRequired,
-    showPlatform: React.PropTypes.func.isRequired,
-    showFramework: React.PropTypes.func.isRequired,
-    installPlatform: React.PropTypes.func.isRequired
+    name: PropTypes.string.isRequired,
+    data: PropTypes.object,
+    loadPlatformData: PropTypes.func.isRequired,
+    showPlatform: PropTypes.func.isRequired,
+    showFramework: PropTypes.func.isRequired,
+    installPlatform: PropTypes.func.isRequired
   }
 
   componentWillMount() {
