@@ -8,6 +8,7 @@
  * the root directory of this source tree.
  */
 
+import PropTypes from 'prop-types';
 import React from 'react';
 import path from 'path';
 
@@ -15,11 +16,11 @@ import path from 'path';
 export default class ProjectsBlock extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(React.PropTypes.string),
-    filterValue: React.PropTypes.string.isRequired,
-    setFilter: React.PropTypes.func.isRequired,
-    openProject: React.PropTypes.func.isRequired,
-    removeProject: React.PropTypes.func.isRequired
+    items: PropTypes.arrayOf(PropTypes.string),
+    filterValue: PropTypes.string,
+    setFilter: PropTypes.func.isRequired,
+    openProject: PropTypes.func.isRequired,
+    removeProject: PropTypes.func.isRequired
   }
 
   render() {
@@ -42,7 +43,10 @@ export default class ProjectsBlock extends React.Component {
         <div className='select-list'>
           <ol className='list-group'>
             { items.map(item => (
-                <ProjectItem item={ item } onOpen={ this.props.openProject } onRemove={ this.props.removeProject } />
+                <ProjectItem item={ item }
+                  key={ item }
+                  onOpen={ this.props.openProject }
+                  onRemove={ this.props.removeProject } />
               )) }
           </ol>
         </div>
@@ -55,9 +59,9 @@ export default class ProjectsBlock extends React.Component {
 class ProjectItem extends React.Component {
 
   static propTypes = {
-    item: React.PropTypes.object.isRequired,
-    onOpen: React.PropTypes.func.isRequired,
-    onRemove: React.PropTypes.func.isRequired
+    item: PropTypes.string.isRequired,
+    onOpen: PropTypes.func.isRequired,
+    onRemove: PropTypes.func.isRequired
   }
 
   componentDidMount() {
