@@ -15,22 +15,23 @@ import { INPUT_FILTER_KEY, getFilter, getVisibleProjects } from '../selectors';
 import { CompositeDisposable } from 'atom';
 import { INPUT_FILTER_DELAY } from '../../config';
 import ProjectsBlock from '../components/projects-block';
+import PropTypes from 'prop-types';
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { lazyUpdateInputValue } from '../../home/actions';
+import { lazyUpdateInputValue } from '../../core/actions';
 
 
 class RecentProjectsBlock extends React.Component {
 
   static propTypes = {
-    items: React.PropTypes.arrayOf(React.PropTypes.string).isRequired,
-    filterValue: React.PropTypes.string.isRequired,
-    setFilter: React.PropTypes.func.isRequired,
-    loadProjects: React.PropTypes.func.isRequired,
-    syncProjects: React.PropTypes.func.isRequired,
-    openProject: React.PropTypes.func.isRequired,
-    removeProject: React.PropTypes.func.isRequired
+    items: PropTypes.arrayOf(PropTypes.string),
+    filterValue: PropTypes.string,
+    setFilter: PropTypes.func.isRequired,
+    loadProjects: PropTypes.func.isRequired,
+    syncProjects: PropTypes.func.isRequired,
+    openProject: PropTypes.func.isRequired,
+    removeProject: PropTypes.func.isRequired
   }
 
   componentWillMount() {
