@@ -8,10 +8,14 @@
  * the root directory of this source tree.
  */
 
-export function getInputValue(state, key) {
-  const data = state.inputValues || {};
-  if (!data.hasOwnProperty(key)) {
-    return null;
+import BaseModal from '../../core/base-modal';
+import FetchTokenModalContainer from '../containers/fetch-token-modal-container';
+
+
+export default class FetchTokenModal extends BaseModal {
+
+  get component() {
+    return FetchTokenModalContainer;
   }
-  return data[key];
+
 }
