@@ -9,7 +9,6 @@
  */
 
 import fs from 'fs-plus';
-import os from 'os';
 import path from 'path';
 
 
@@ -102,16 +101,23 @@ export const ATOM_CONFIG = {
         type: 'string',
         default: '',
         order: 30
-      }
+      },
+      checkAuthOnStartup: {
+        title: 'Check PIO Account authentication status on IDE startup',
+        description: 'If the user is not logged in, a login form will be displayed.',
+        type: 'boolean',
+        default: true,
+        order: 40
+      },
     }
   }
 };
 
-export const IS_WINDOWS = Boolean(os.platform().includes('win32'));
+export const IS_WINDOWS = process.platform.startsWith('win');
 export const PKG_BASE_DIR = path.resolve(path.dirname(__filename), '..');
 export const CACHE_DIR = path.join(PKG_BASE_DIR, '.cache');
-export const PIO_HOME_DIR = process.env.PLATFOMRIO_HOME_DIR ? process.env.PLATFORMIO_HOME_DIR : path.join(fs.getHomeDirectory(), '.platformio');
-export const ENV_DIR = _get_env_dir(path.join(PIO_HOME_DIR, 'penv'));
+export const PIO_HOME_DIR = _getPioHomeDir(process.env.PLATFORMIO_HOME_DIR || path.join(fs.getHomeDirectory() || '~', '.platformio'));
+export const ENV_DIR = path.join(PIO_HOME_DIR, 'penv');
 export const ENV_BIN_DIR = path.join(ENV_DIR, IS_WINDOWS ? 'Scripts' : 'bin');
 
 export const ATOM_DEPENDENCIES = {
@@ -126,8 +132,11 @@ export const ATOM_DEPENDENCIES = {
   'platformio-ide-terminal': {
     requirements: '>=2.5.0'
   },
+  'platformio-ide-debugger': {
+    requirements: '>=1.2.0'
+  },
   'autocomplete-clang': {
-    requirements: '>=0.8.9'
+    requirements: '>=0.11.3'
   },
   'linter-ui-default': {
     requirements: '>=1.1.0'
@@ -158,10 +167,22 @@ export const ATOM_DEPENDENCIES = {
 export const ATOM_CONFLICTED_DEPENDENCIES = {
   'build-platformio': {
     requirements: '*'
+  },
+  'busy-signal': {
+    requirements: '*'
+  },
+  'dbg': {
+    requirements: '*'
+  },
+  'dbg-gdb': {
+    requirements: '*'
+  },
+  'output-panel': {
+    requirements: '*'
   }
 };
 
-export const PIO_CORE_MIN_VERSION = '3.3.0';
+export const PIO_CORE_MIN_VERSION = '3.4.0-b.1';
 export const DEFAULT_PIO_ARGS = ['-f', '-c', 'atom'];
 export const TERMINAL_REOPEN_DELAY = 1000; // ms, dalay before serial monitor restore
 export const AUTO_REBUILD_DELAY = 3000; // ms
@@ -169,22 +190,21 @@ export const INPUT_FILTER_DELAY = 200; // ms, dalay before filtering projects, l
 
 export const PLATFORMIO_API_ENDPOINT = 'http://api.platformio.org';
 
-
-function _get_env_dir(defaultEnvDir) {
+function _getPioHomeDir(pioHomeDir) {
   if (IS_WINDOWS) {
-    // Put the env directory to the root of the current local disk when
-    // default path contains non-ASCII characters. Virtualenv will fail to
-    for (const char of defaultEnvDir) {
+    // Make sure that all path characters have valid ASCII codes.
+    for (const char of pioHomeDir) {
       if (char.charCodeAt(0) > 127) {
-        const defaultEnvDirFormat = path.parse(defaultEnvDir);
+        // If they don't, put the pio home directory into the root of the disk.
+        const homeDirPathFormat = path.parse(pioHomeDir);
         return path.format({
-          root: defaultEnvDirFormat.root,
-          dir: defaultEnvDirFormat.root,
-          base: '.pioidepenv',
-          name: '.pioidepenv',
+          root: homeDirPathFormat.root,
+          dir: homeDirPathFormat.root,
+          base: '.platformio',
+          name: '.platformio',
         });
       }
     }
   }
-  return defaultEnvDir;
+  return pioHomeDir;
 }
