@@ -8,6 +8,7 @@
  * the root directory of this source tree.
  */
 
+import { CompositeDisposable } from 'atom';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
@@ -16,12 +17,21 @@ export default class BaseModal {
 
   constructor(props) {
     this.props = props || {};
+
     this._promise = new Promise((resolve, reject) => {
       this._onResolve = resolve;
       this._onReject = reject;
     });
     this._panel = null;
     this._element = document.createElement('div');
+    this._element.classList.add('pio-modal');
+    this._element.classList.add('pio-native-key-bindings-tab-fix');
+    this._element.classList.add('native-key-bindings');
+
+    this.subscriptions = new CompositeDisposable();
+    this.subscriptions.add(atom.commands.add('atom-workspace', {
+      'core:cancel': () => this.reject(null)
+    }));
   }
 
   get component() {
@@ -41,10 +51,7 @@ export default class BaseModal {
 
   renderComponent() {
     ReactDOM.render(
-      <this.component
-        { ...this.props }
-        onResolve={ ::this.resolve }
-        onReject={ ::this.reject } />,
+      <this.component { ...this.props } onResolve={ ::this.resolve } onReject={ ::this.reject } />,
       this._element
     );
   }
@@ -67,6 +74,7 @@ export default class BaseModal {
   onDidOpen() {}
 
   destroy() {
+    this.subscriptions.dispose();
     this._promise = null;
     if (this._panel) {
       this._panel.destroy();
