@@ -11,6 +11,7 @@
 import * as ActionTypes from './actions';
 
 import { combineReducers } from 'redux';
+import { copyWithoutMatchingKeys } from './helpers';
 
 
 function entities(state = {}, action) {
@@ -20,15 +21,8 @@ function entities(state = {}, action) {
         [action.key]: action.data
       });
 
-    case ActionTypes.DELETE_ENTITY: {
-      const newState = Object.assign({}, state);
-      Object.keys(newState).forEach(key => {
-        if (action.re.test(key)) {
-          delete newState[key];
-        }
-      });
-      return newState;
-    }
+    case ActionTypes.DELETE_ENTITY:
+      return copyWithoutMatchingKeys(state, action.re);
 
     case ActionTypes.RESET_STORE:
       return {};
@@ -49,8 +43,40 @@ function inputValues(state = {}, action) {
   return state;
 }
 
+function flags(state = {}, action) {
+  switch (action.type) {
+    case ActionTypes.UPDATE_FLAG_VALUE:
+      return Object.assign({}, state, {
+        [action.key]: action.value
+      });
+
+    case ActionTypes.RESET_STORE:
+      return {};
+  }
+  return state;
+}
+
+function errors(state = {}, action) {
+  switch (action.type) {
+    case ActionTypes.UPDATE_ERROR:
+      return {
+        ...state,
+        [action.key]: action.value,
+      };
+
+    case ActionTypes.DELETE_ERROR:
+      return copyWithoutMatchingKeys(state, action.re);
+
+    case ActionTypes.RESET_STORE:
+      return {};
+  }
+  return state;
+}
+
 const rootReducer = combineReducers({
   entities,
+  errors,
+  flags,
   inputValues
 });
 
