@@ -55,10 +55,10 @@ export default class WelcomePage extends React.Component {
             <PioVersions />
           </div>
           <div className='col-xs quick-links'>
-            <h2>Quick access</h2>
+            <h2>Quick Access</h2>
             <ul className='list-group'>
               { this.getQuickLinks().map(item => (
-                  <li className='list-item'>
+                  <li className='list-item' key={ item.text }>
                     <a className={ 'btn btn-lg icon icon-' + item.icon } onClick={ item.callback }>
                       { item.text }
                     </a>
@@ -106,12 +106,12 @@ export default class WelcomePage extends React.Component {
               ·
             </li>
             <li>
-              <span className='icon icon-bug'></span> <a href='https://github.com/platformio/platformio/issues'>Report an issue</a>
+              <span className='icon icon-bug'></span> <a href='https://github.com/platformio/platformio/issues'>Report an Issue</a>
             </li>
           </ul>
           <hr />
         </div>
-        <h2 className='block section-heading icon icon-file-submodule'>Recent projects</h2>
+        <h2 className='block section-heading icon icon-file-submodule'>Recent Projects</h2>
         <RecentProjectsBlock />
         <hr />
         <p className='text-center'>
