@@ -14,17 +14,13 @@ import { Provider } from 'react-redux';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Telemetry from '../telemetry';
-import configureStore from './store';
-import { resetStore } from './actions';
-import rootSaga from './sagas';
+import { getStore } from '../core/store';
 
 
 export default class HomeIndex {
 
   constructor(uri) {
     this.uri = uri;
-    this.store = configureStore();
-    this.store.runSaga(rootSaga);
     this.element = document.createElement('div');
 
     let startLocation = this.uri.replace('platformio://home', '');
@@ -33,7 +29,7 @@ export default class HomeIndex {
     }
 
     ReactDOM.render(
-      <Provider store={ this.store }>
+      <Provider store={ getStore() }>
         <MemoryRouter initialEntries={ [startLocation] }>
           <HomeApp />
         </MemoryRouter>
@@ -41,14 +37,11 @@ export default class HomeIndex {
       this.element
     );
 
-    Telemetry.hitScreenView('home' + startLocation);
+    Telemetry.hitScreenView(startLocation);
   }
 
   destroy() {
     ReactDOM.unmountComponentAtNode(this.element);
-    this.store.dispatch(resetStore());
-    this.store.close();
-    this.store = null;
   }
 
   getURI() {
