@@ -10,24 +10,11 @@
 
 /* eslint-disable no-constant-condition */
 
-import * as actions from './actions';
-
-import { call, fork, put, takeEvery, takeLatest } from 'redux-saga/effects';
+import { fork, takeEvery } from 'redux-saga/effects';
 
 import Telemetry from '../telemetry';
-import { asyncDelay } from '../utils';
-import librarySagas from '../library/sagas';
 import path from 'path';
-import platformSagas from '../platform/sagas';
-import projectSagas from '../project/sagas';
 
-
-function* watchLazyUpdateInputValue() {
-  yield takeLatest(actions.LAZY_UPDATE_INPUT_VALUE, function* ({ key, value, delay}) {
-    yield call(asyncDelay, delay);
-    yield put(actions.updateInputValue(key, value));
-  });
-}
 
 function* watchEventActions() {
   const re = new RegExp('^(install|uninstall|update|open|remove)_(library|platform|project)', 'i');
@@ -46,12 +33,6 @@ function* watchEventActions() {
   });
 }
 
-export default function* root() {
-  yield [
-    watchLazyUpdateInputValue,
-    watchEventActions,
-    ...librarySagas,
-    ...projectSagas,
-    ...platformSagas
-  ].map(item => fork(item));
-}
+export default [
+  watchEventActions
+];
