@@ -8,7 +8,8 @@
  * the root directory of this source tree.
  */
 
-import {handleLibChanges, initializeProject, installPlatformsForBoards} from '../init/command';
+import { initializeProject, installPlatformsForBoards } from '../init/command';
+
 import {ImportArduinoProjectView} from './view';
 import {Transform} from 'stream';
 import fs from 'fs-extra';
@@ -50,7 +51,6 @@ export function command() {
     chain = chain.then(addIncludeToInoFiles);
 
     chain = chain.then(() => atom.project.addPath(projectPath));
-    chain = chain.then(() => handleLibChanges([projectPath]));
 
     chain = chain.then(() => {
       let notifyMessage = 'The next files/directories were created in "' + projectPath + '"\n' +
