@@ -25,12 +25,41 @@ export function getCacheDir() {
   return config.CACHE_DIR;
 }
 
-export async function download(source, target) {
-  // check if file is already downloaded
+export async function download(source, target, retries = 3) {
+  const contentLength = await getContentLength(source);
+
+  if (fileExistsAndSizeMatches(target, contentLength)) {
+    return target;
+  }
+
+  let lastError = null;
+
+  while (retries >= 0) {
+    try {
+      await _download(source, target);
+      if (fileExistsAndSizeMatches(target, contentLength)) {
+        return target;
+      }
+    } catch (error) {
+      lastError = error;
+      console.error(error);
+    }
+
+    retries--;
+  }
+
+  if (lastError) {
+    throw lastError;
+  } else {
+    throw new Error(`Failed to download file ${path.basename(target)}`);
+  }
+}
+
+
+function fileExistsAndSizeMatches(target, contentLength) {
   if (fs.isFileSync(target)) {
-    const contentLength = await getContentLength(source);
-    if (contentLength > 0 && fs.getSizeSync(target)) {
-      return target;
+    if (contentLength > 0 && contentLength == fs.getSizeSync(target)) {
+      return true;
     }
     try {
       fs.removeSync(target);
@@ -38,7 +67,11 @@ export async function download(source, target) {
       console.error(err);
     }
   }
+  return false;
+}
 
+
+async function _download(source, target) {
   return new Promise((resolve, reject) => {
     utils.runAPMCommand(
       ['config', 'get', 'https-proxy'],
