@@ -107,7 +107,7 @@ function upgradeIDE(branch) {
           return resolve(true);
         } else {
           utils.notifyError(
-            'Problem has beenn occured while uprading PlatformIO IDE.',
+            'Problem occured while uprading PlatformIO IDE.',
             new Error(stderr)
           );
           return reject(stderr);
@@ -125,7 +125,7 @@ export async function reinstallIDE(useDevelop) {
     await new Promise((resolve, reject) => {
       utils.runAPMCommand(
         ['uninstall', 'platformio-ide'],
-        (code, stdout, stderr) => code === 0 ? resolve(stdout) : reject(stderr),
+        (code, stdout, stderr) => code === 0 ? resolve(stdout) : reject(stdout.toString() + stderr.toString()),
         {
           busyTitle: 'Uninstalling PlatformIO IDE'
         }
@@ -138,7 +138,7 @@ export async function reinstallIDE(useDevelop) {
           '--production',
           useDevelop ? 'platformio/platformio-atom-ide' : 'platformio-ide'
         ],
-        (code, stdout, stderr) => code === 0 ? resolve(stdout) : reject(stderr),
+        (code, stdout, stderr) => code === 0 ? resolve(stdout) : reject(stdout.toString() + stderr.toString()),
         {
           busyTitle: 'Installing PlatformIO IDE'
         }
@@ -147,7 +147,7 @@ export async function reinstallIDE(useDevelop) {
 
   } catch (err) {
     utils.notifyError(
-      'Problem has beenn occured while installing PlatformIO IDE.',
+      'Problem has been occured while installing PlatformIO IDE.',
       new Error(err)
     );
     return false;
