@@ -17,6 +17,8 @@ import semver from 'semver';
 
 export default class AtomDependenciesStage extends BaseStage {
 
+  static MAX_FAILURES_COUNT = 3;
+
   get name() {
     return 'Atom dependencies';
   }
@@ -67,24 +69,32 @@ export default class AtomDependenciesStage extends BaseStage {
     );
   }
 
-  installMissed() {
-    return new Promise((resolve, reject) => {
-      let missed = this.getMissed();
-      if (missed.length == 0) {
-        return resolve(true);
+  async installMissed() {
+    let missed = this.getMissed();
+    if (missed.length == 0) {
+      return true;
+    }
+    missed = missed.map(name => {
+      if (!config.ATOM_DEPENDENCIES[name].forceVersion) {
+        return name;
       }
-      missed = missed.map(name => {
-        if (!config.ATOM_DEPENDENCIES[name].forceVersion) {
-          return name;
-        }
-        return `${name}@${config.ATOM_DEPENDENCIES[name].forceVersion}`;
-      });
+      return `${name}@${config.ATOM_DEPENDENCIES[name].forceVersion}`;
+    });
+
+    try {
+      await this._installMissedByNames(missed);
+    } catch (err) {
+      await this._installMissedOneByOne(missed);
+    }
+  }
+
+  _installMissedByNames(missed) {
+    return new Promise((resolve, reject) => {
       utils.runAPMCommand(
         ['install', ...missed, '--production', '--compatible'],
         (code, stdout, stderr) => {
           if (code !== 0) {
             const error = new Error(stderr);
-            utils.notifyError(`Install missed packages: ${missed}`, error);
             return reject(error);
           } else {
             console.debug(stdout);
@@ -93,6 +103,24 @@ export default class AtomDependenciesStage extends BaseStage {
         }
       );
     });
+  }
+
+  async _installMissedOneByOne(missed) {
+    let failuresCount = 0;
+    while (missed.length > 0) {
+      const name = missed.shift();
+      try {
+        await this._installMissedByNames([name]);
+      } catch (err) {
+        failuresCount++;
+        missed.push(name);
+
+        if (failuresCount >= this.MAX_FAILURES_COUNT) {
+          utils.notifyError(`Failed to install missed packages: ${missed}`, err);
+          throw new Error(err.toString());
+        }
+      }
+    }
   }
 
   uninstallIncompatible() {
