@@ -18,7 +18,9 @@ export default class ConfigurationStage extends BaseStage {
     'platformio-ide:donate:auto-popup-enabled',
     'platformio-ide:donate:conter',
     'platformio-ide:clang-checked',
-    'platformio-ide.defaultToolbarPositionHasBeenSet'
+    'platformio-ide.defaultToolbarPositionHasBeenSet',
+    'platformio-ide:pioplus-username',
+    'platformio-ide:pioplus-logged-in',
   ];
 
   static OBSOLATE_CONFIG_ITEMS = [
