@@ -23,7 +23,7 @@ export default class PlatformIOCoreStage extends BaseStage {
   static UPGRADE_PIOCORE_TIMEOUT =86400 * 3 * 1000; // 3 days
 
   static pythonVersion = '2.7.13';
-  static vitrualenvUrl = 'https://pypi.python.org/packages/d4/0c/9840c08189e030873387a73b90ada981885010dd9aea134d6de30cd24cb8/virtualenv-15.1.0.tar.gz';
+  static vitrualenvUrl = 'https://pypi.python.org/packages/source/v/virtualenv/virtualenv-14.0.6.tar.gz';
 
   constructor() {
     super(...arguments);
@@ -86,38 +86,19 @@ export default class PlatformIOCoreStage extends BaseStage {
     // https://www.python.org/ftp/python/2.7.13/python-2.7.13.amd64.msi
     const pythonArch = process.arch === 'x64' ? '.amd64' : '';
     const msiUrl = `https://www.python.org/ftp/python/${PlatformIOCoreStage.pythonVersion}/python-${PlatformIOCoreStage.pythonVersion}${pythonArch}.msi`;
+    const msiInstaller = await download(
+      msiUrl,
+      path.join(getCacheDir(), path.basename(msiUrl))
+    );
     const targetDir = path.join(config.PIO_HOME_DIR, 'python27');
     const pythonPath = path.join(targetDir, 'python.exe');
 
     if (!fs.isFileSync(pythonPath)) {
-      const msiInstaller = await download(
-        msiUrl,
-        path.join(getCacheDir(), path.basename(msiUrl))
-      );
-      const logFile = path.join(getCacheDir(), 'python27msi.log');
-      await new Promise((resolve, reject) => {
-        utils.runCommand(
-          'msiexec.exe',
-          ['/i', msiInstaller, '/qn', '/li', logFile, `TARGETDIR=${targetDir}`],
-          (code, stdout, stderr) => {
-            if (code === 0) {
-              return resolve(stdout);
-            } else {
-              if (fs.isFileSync(logFile)) {
-                stderr = fs.readFileSync(logFile).toString();
-              }
-              return reject(`MSI Python2.7: ${stderr}`);
-            }
-          },
-          {
-            spawnOptions: {
-              shell: true
-            }
-          }
-        );
-      });
-      if (!fs.isFileSync(pythonPath)) {
-        throw new Error('Could not install Python 2.7 using MSI');
+      try {
+        await this.installPythonFromWindowsMSI(msiInstaller, targetDir);
+      } catch (err) {
+        console.error(err);
+        await this.installPythonFromWindowsMSI(msiInstaller, targetDir, true);
       }
     }
 
@@ -133,6 +114,34 @@ export default class PlatformIOCoreStage extends BaseStage {
         () => resolve(pythonPath)
       );
     });
+  }
+
+  async installPythonFromWindowsMSI(msiInstaller, targetDir, administrative = false) {
+    const logFile = path.join(getCacheDir(), 'python27msi.log');
+    await new Promise((resolve, reject) => {
+      utils.runCommand(
+        'msiexec.exe',
+        [administrative ? '/a' : '/i', msiInstaller, '/qn', '/li', logFile, `TARGETDIR=${targetDir}`],
+        (code, stdout, stderr) => {
+          if (code === 0) {
+            return resolve(stdout);
+          } else {
+            if (fs.isFileSync(logFile)) {
+              stderr = fs.readFileSync(logFile).toString();
+            }
+            return reject(`MSI Python2.7: ${stderr}`);
+          }
+        },
+        {
+          spawnOptions: {
+            shell: true
+          }
+        }
+      );
+    });
+    if (!fs.isFileSync(path.join(targetDir, 'python.exe'))) {
+      throw new Error('Could not install Python 2.7 using MSI');
+    }
   }
 
   chooseCustomPythonDirs() {
