@@ -11,7 +11,7 @@
 import { LibraryStorage, filterStorageItems, getLibraryStorages } from './storage';
 import { expandFrameworksOrPlatforms, getRegistryFrameworks, getRegistryPlatforms } from '../platform/selectors';
 
-import { getInputValue } from '../home/selectors';
+import { getInputValue } from '../core/selectors';
 
 
 // Data Filters
