@@ -15,14 +15,18 @@ import * as utils from './utils';
 
 import { TerminalConsumer, runCmdsInTerminal } from './services/terminal';
 import { checkIDEUpdates, reinstallIDE } from './installer/ide';
+import { destroyStore, initStore, serializeStore } from './core/store';
 import { getActivePioProject, synchronizeRecentProjects } from './project/helpers';
 
 import BuildProvider from './services/build';
 import { BusyConsumer } from './services/busy';
 import { CompositeDisposable } from 'atom';
+import { DebuggerConsumer } from './services/debugger';
 import HomeIndex from './home/index';
 import InstallationManager from './installer/manager';
+import ProjectObserver from './project/observer';
 import { ToolbarConsumer } from './services/toolbar';
+import { checkAuthStatusUnlessDisabled } from './account/helpers';
 import { command as donateCommand } from './donate/command';
 import { command as importArduinoIDEProject } from './import-arduino-project/command';
 import { openPIOHome } from './home/helpers';
@@ -44,15 +48,24 @@ class PlatformIOIDEPackage {
     this.consumePlatformioIDETerminal = TerminalConsumer;
     this.consumeBusy = BusyConsumer;
     this.consumeToolbar = ToolbarConsumer;
+    this.consumeDebugger = DebuggerConsumer;
 
     // misc
     this._reinstallIDETimer = null;
   }
 
-  activate() {
+  activate(state) {
+    this.projectObservable = new ProjectObserver();
+
     window.addEventListener('error', maintenance.handleUncaughtErrors);
     maintenance.updateOSEnviron();
+
+    if (state && state.hasOwnProperty('store')) {
+      initStore(state.store);
+    }
+
     this.startInstaller()
+      .then(() => checkAuthStatusUnlessDisabled())
       .then(() => this.setupCommands())
       .then(() => donateCommand(true))
       .then(() => synchronizeRecentProjects(atom.project.getPaths()))
@@ -90,18 +103,6 @@ class PlatformIOIDEPackage {
       'platformio-ide:maintenance.open-terminal': () => runCmdsInTerminal(['pio --help']),
       'platformio-ide:maintenance.serial-monitor': () => serialMonitor(),
       'platformio-ide:maintenance.serial-ports': () => runCmdsInTerminal(['pio device list']),
-
-      'platformio-ide:maintenance.rebuild-index': () => {
-        const p = getActivePioProject();
-        if (!p) {
-          atom.notifications.addWarning('Can not find active PlatformIO project.', {
-            detail: 'Make sure that an opened project you are trying to rebuid is ' +
-              'a PlatformIO project (e.g., contains `platformio.ini`).'
-          });
-          return;
-        }
-        return init.intendToPerformIndexRebuild(p);
-      },
       'platformio-ide:maintenance.install-commands': () => maintenance.installCommands(),
       'platformio-ide:maintenance.update-platformio': () => runCmdsInTerminal(['platformio update']),
       'platformio-ide:maintenance.upgrade-platformio': () => runCmdsInTerminal(['platformio upgrade']),
@@ -172,7 +173,7 @@ class PlatformIOIDEPackage {
       })
     );
 
-    for (const target of ['build', 'upload', 'program', 'clean', 'test']) {
+    for (const target of ['build', 'upload', 'program', 'clean', 'test', 'debug']) {
       this.subscriptions.add(atom.commands.add(
         'atom-workspace',
         `platformio-ide:target:${target}`,
@@ -227,16 +228,6 @@ class PlatformIOIDEPackage {
     this.subscriptions.add(atom.project.onDidChangePaths((projectPaths) => {
       synchronizeRecentProjects(projectPaths);
       utils.runAtomCommand('tree-view:show');
-      init.ensureProjectsInited(projectPaths);
-      init.handleLibChanges(projectPaths);
-    }));
-
-    this.subscriptions.add(atom.config.observe('platformio-ide.autoRebuildAutocompleteIndex', (enabled) => {
-      if (enabled) {
-        init.handleLibChanges(atom.project.getPaths());
-      } else {
-        init.clearLibChangeWatchers();
-      }
     }));
 
     this.subscriptions.add(
@@ -265,15 +256,21 @@ class PlatformIOIDEPackage {
     doHighlight();
   }
 
+  serialize() {
+    return {
+      store: serializeStore()
+    };
+  }
+
   deactivate() {
+    destroyStore();
     this.subscriptions.dispose();
     if (this.highlightSubscriptions) {
       this.highlightSubscriptions.dispose();
     }
-    init.clearLibChangeWatchers();
-    init.cleanMiscCache();  // @FIXME
     utils.cleanMiscCache(); // @FIXME
     window.removeEventListener('error', maintenance.handleUncaughtErrors);
+    this.projectObservable.destroy();
   }
 
 }
