@@ -157,13 +157,9 @@ export function highlightActiveProject(isEnabled = true, retries = 3) {
 
   dirs.forEach((dir) => {
     dir.classList.remove('pio-active-directory');
-    if (!dir.dataset.pioRealpath) {
+    if (!dir.dataset.pioProjectPath) {
       const span = dir.querySelector('.header > span[data-path]');
-      try {
-        dir.dataset.pioRealpath = fs.realpathSync(span.dataset.path);
-      } catch (e) {
-        dir.dataset.pioRealpath = span.dataset.path;
-      }
+      dir.dataset.pioProjectPath = span.dataset.path;
     }
   });
   if (dirs.length < 2 || !isEnabled || !p) {
@@ -172,7 +168,7 @@ export function highlightActiveProject(isEnabled = true, retries = 3) {
 
   let done = false;
   for (const dir of dirs) {
-    if (dir.dataset.pioRealpath === p.toString()) {
+    if (dir.dataset.pioProjectPath === p.toString()) {
       dir.classList.add('pio-active-directory');
       done = true;
       break;
