@@ -14,7 +14,7 @@ import * as actions from './actions';
 import * as selectors from './selectors';
 
 import { call, put, select, take, takeEvery, takeLatest } from 'redux-saga/effects';
-import { deleteEntity, updateEntity } from '../home/actions';
+import { deleteEntity, updateEntity } from '../core/actions';
 import { fetchBoards, fetchRegistryPackages, runPlatformCommand } from './helpers';
 
 
