@@ -11,7 +11,7 @@
 import * as utils from '../utils';
 
 import fuzzaldrin from 'fuzzaldrin-plus';
-import { getInputValue } from '../home/selectors';
+import { getInputValue } from '../core/selectors';
 
 
 // Data Filters
