@@ -9,16 +9,17 @@
  */
 
 import * as config from '../config';
-import {ProjectExamplesView} from './view';
+import { ProjectExamplesView } from './view';
 import fs from 'fs-plus';
 import path from 'path';
-import {rebuildIndex} from '../init/command';
 
 
 export async function command() {
   const projects = await getProjects(path.join(config.PIO_HOME_DIR, 'project-examples'));
   const view = new ProjectExamplesView(projects);
-  const panel = atom.workspace.addModalPanel({item: view.getElement()});
+  const panel = atom.workspace.addModalPanel({
+    item: view.getElement()
+  });
   let canceled = false;
 
   view.handlePrepare = async function() {
@@ -32,7 +33,6 @@ export async function command() {
       step += 1;
       if (!canceled) {
         view.setStatus(`Processing project "${path.basename(projectPath)}"`);
-        await rebuildIndex(projectPath);
         atom.project.addPath(projectPath);
         processedPaths.push(projectPath);
       }
@@ -55,7 +55,7 @@ async function getProjects(examplesRoot) {
   const queue = [examplesRoot];
   const projects = {};
   while (queue.length > 0) {
-    const dirPath = queue.splice(0, 1)[0];  // take the first element from the queue
+    const dirPath = queue.splice(0, 1)[0]; // take the first element from the queue
     if (!dirPath) {
       continue;
     }
@@ -70,7 +70,7 @@ async function getProjects(examplesRoot) {
         if (fs.isDirectorySync(fullPath) && file !== 'ide') {
           queue.push(fullPath);
         }
-      } catch(e) {
+      } catch (e) {
         continue;
       }
     }
