@@ -17,7 +17,7 @@ import { getRecentProjects, removeRecentProject, synchronizeRecentProjects } fro
 
 import { asyncDelay } from '../utils';
 import { getProjects } from './selectors';
-import { updateEntity } from '../home/actions';
+import { updateEntity } from '../core/actions';
 
 
 function* reloadProjects() {
