@@ -9,7 +9,7 @@
  */
 
 import fuzzaldrin from 'fuzzaldrin-plus';
-import { getInputValue } from '../home/selectors';
+import { getInputValue } from '../core/selectors';
 
 
 export const INPUT_FILTER_KEY = 'projectFilter';
