@@ -11,13 +11,15 @@
 import * as utils from '../utils';
 
 import { getTerminalViews, runCmdsInTerminal } from './terminal';
+
 import { TERMINAL_REOPEN_DELAY } from '../config';
 import Telemetry from '../telemetry';
+import { debugProject } from './debugger.js';
 import fs from 'fs';
 import ini from 'ini';
 import { isPioProject } from '../project/helpers';
+import openAuthModal from '../account/helpers';
 import path from 'path';
-
 
 export default class BuildProvider {
 
@@ -37,11 +39,6 @@ export default class BuildProvider {
       keymap: 'ctrl-alt-c',
     },
     {
-      name: 'Test',
-      args: ['test'],
-      keymap: 'ctrl-alt-shift-t',
-    },
-    {
       name: 'Upload',
       args: ['run', '--target', 'upload'],
       keymap: 'ctrl-alt-u',
@@ -53,15 +50,21 @@ export default class BuildProvider {
     {
       name: 'Upload SPIFFS image',
       args: ['run', '--target', 'uploadfs'],
-    }
+    },
+    {
+      name: 'Test',
+      args: ['test'],
+      keymap: 'ctrl-alt-shift-t',
+    },
+    {
+      name: 'Debug',
+      args: ['debug'],
+      keymap: 'ctrl-alt-shift-t',
+    },
   ];
 
   constructor(cwd) {
-    try {
-      this.projectDir = fs.realpathSync(cwd);
-    } catch (e) {
-      this.projectDir = cwd;
-    }
+    this.projectDir = cwd;
   }
 
   getNiceName() {
@@ -256,17 +259,30 @@ class RunTarget {
   }
 
   get postBuild() {
-    return buildOutcome => {
-      if (!buildOutcome || !this._terminalsToRestore) {
+    return async (buildOutcome, stdout, stderr) => {
+      if (stderr && stderr.toString().includes('pio account login')) {
+        openAuthModal();
+      }
+
+      if (!buildOutcome) {
         return;
       }
-      setTimeout(() => {
-        this._terminalsToRestore.forEach(item => {
-          runCmdsInTerminal([item]);
-        });
-        this._terminalsToRestore = null;
-      }, TERMINAL_REOPEN_DELAY);
+
+      if (this._coreTargetName === 'debug') {
+        debugProject(this._projectDir, this._coreEnvName);
+      }
+
+      if (this._terminalsToRestore) {
+        setTimeout(() => {
+          this._terminalsToRestore.forEach(item => {
+            runCmdsInTerminal([item]);
+          });
+          this._terminalsToRestore = null;
+        }, TERMINAL_REOPEN_DELAY);
+      }
     };
   }
+
+
 
 }
