@@ -93,18 +93,20 @@ export function runCommand(cmd, args, callback, options = {}) {
     }
   }
 
-  return new BufferedProcess({
+  const p = new BufferedProcess({
     command: cmd,
     args: args,
     options: options.spawnOptions,
     stdout: (line) => outputLines.push(line),
     stderr: (line) => errorLines.push(line),
     exit: (code) => onExit(code)
-  }).onWillThrowError(errorObject => {
+  });
+  p.onWillThrowError(errorObject => {
     errorObject.handle();
     errorLines.push(errorObject.error.toString());
     onExit(-1);
   });
+  return p;
 }
 
 export function runPIOCommand(args, callback, options = {}) {
